+++ sycl/source/builtins/common_functions.cpp
@@ -17,6 +17,7 @@
 #include <cmath>
 
 #include <sycl/detail/builtins/builtins.hpp>
+#include <sycl/vector.hpp>
 
 #include "host_helper_macros.hpp"
