+++ sycl/source/builtins/math_functions.cpp
@@ -15,6 +15,7 @@
 #include <cmath>
 
 #include <sycl/detail/builtins/builtins.hpp>
+#include <sycl/vector.hpp>
 
 #include "host_helper_macros.hpp"
