+++ sycl/test-e2e/DeviceLib/built-ins/fast-math-flag.cpp
@@ -5,6 +5,7 @@
 #include <sycl/detail/core.hpp>
 
 #include <sycl/builtins.hpp>
+#include <sycl/vector.hpp>
 
 using namespace sycl;
