+++ sycl/test-e2e/DeviceLib/built-ins/vector_math.cpp
@@ -4,6 +4,7 @@
 #include <sycl/detail/core.hpp>
 
 #include <sycl/builtins.hpp>
+#include <sycl/vector.hpp>
 
 #include <array>
 #include <cassert>
