+++ sycl/test-e2e/USM/math.cpp
@@ -3,6 +3,7 @@
 
 #include <sycl/detail/core.hpp>
 #include <sycl/usm.hpp>
+#include <sycl/vector.hpp>
 
 #include <array>
 #include <cassert>
