+++ sycl/test/include_deps/sycl_khr_includes_images.hpp.cpp
@@ -104,8 +104,10 @@
 // CHECK-NEXT: event.hpp
 // CHECK-NEXT: sampler.hpp
 // CHECK-NEXT: vector.hpp
-// CHECK-NEXT: detail/named_swizzles_mixin.hpp
 // CHECK-NEXT: detail/vector_arith.hpp
+// CHECK-NEXT: detail/vector_traits.hpp
+// CHECK-NEXT: detail/vector_core.hpp
+// CHECK-NEXT: detail/named_swizzles_mixin.hpp
 // CHECK-NEXT: detail/memcpy.hpp
 // CHECK-NEXT: detail/image_ocl_types.hpp
 // CHECK-NEXT: __spirv/spirv_ops.hpp
