+++ sycl/test/include_deps/sycl_khr_includes_reduction.hpp.cpp
@@ -103,10 +103,10 @@
 // CHECK-NEXT: detail/memcpy.hpp
 // CHECK-NEXT: builtins.hpp
 // CHECK-NEXT: detail/builtins/builtins.hpp
-// CHECK-NEXT: marray.hpp
-// CHECK-NEXT: vector.hpp
+// CHECK-NEXT: detail/vector_core.hpp
 // CHECK-NEXT: detail/named_swizzles_mixin.hpp
-// CHECK-NEXT: detail/vector_arith.hpp
+// CHECK-NEXT: detail/vector_traits.hpp
+// CHECK-NEXT: marray.hpp
 // CHECK-NEXT: detail/builtins/common_functions.inc
 // CHECK-NEXT: detail/builtins/helper_macros.hpp
 // CHECK-NEXT: detail/builtins/geometric_functions.inc
@@ -133,6 +133,8 @@
 // CHECK-NEXT: group_algorithm.hpp
 // CHECK-NEXT: functional.hpp
 // CHECK-NEXT: known_identity.hpp
+// CHECK-NEXT: vector.hpp
+// CHECK-NEXT: detail/vector_arith.hpp
 // CHECK-NEXT: nd_item.hpp
 // CHECK-NEXT: ext/oneapi/functional.hpp
 // CHECK-NEXT: handler.hpp
