+++ sycl/test/include_deps/sycl_khr_includes_stream.hpp.cpp
@@ -97,11 +97,11 @@
 // CHECK-NEXT: __spirv/spirv_ops.hpp
 // CHECK-NEXT: builtins.hpp
 // CHECK-NEXT: detail/builtins/builtins.hpp
-// CHECK-NEXT: marray.hpp
-// CHECK-NEXT: vector.hpp
+// CHECK-NEXT: detail/vector_core.hpp
 // CHECK-NEXT: detail/named_swizzles_mixin.hpp
-// CHECK-NEXT: detail/vector_arith.hpp
+// CHECK-NEXT: detail/vector_traits.hpp
 // CHECK-NEXT: detail/memcpy.hpp
+// CHECK-NEXT: marray.hpp
 // CHECK-NEXT: detail/builtins/common_functions.inc
 // CHECK-NEXT: detail/builtins/helper_macros.hpp
 // CHECK-NEXT: detail/builtins/geometric_functions.inc
