+++ sycl/test/regression/builtins_multi_ptr_include_order.cpp
@@ -23,6 +23,7 @@ int main() {
 #elif defined(TEST_BUILTINS_FIRST)
 #include <sycl/builtins.hpp>
 #include <sycl/multi_ptr.hpp>
+#include <sycl/vector.hpp>
 
 SYCL_EXTERNAL void
 testScalar(sycl::multi_ptr<float, sycl::access::address_space::global_space,
@@ -47,6 +48,7 @@ int main() { return 0; }
 // clang-format off
 #include <sycl/multi_ptr.hpp>
 #include <sycl/builtins.hpp>
+#include <sycl/vector.hpp>
 // clang-format on
 
 SYCL_EXTERNAL void
